Eligibility warnings: give the stuck transfer its own notice component
Overloading HardBlockingNotice put a ternary on every prop and gave one
component two messages, on a component six other surfaces render. A separate
notice keeps hold-list.tsx untouched, and matches the shape TransferFailureNotice
already uses for a failed transfer.
